+++ CenPositionMinimum/Robot.jl
@@ -16,7 +16,7 @@ function initialize_model(
     # initialize model
     space = GridSpace(extent, periodic = false, metric = :euclidean)
     scheduler = Schedulers.fastest  # they are executed 
-    gridmap = MMatrix{extent[1],extent[2]}(Int8.(-2*ones(Int8, extent)))
+    gridmap = MMatrix{extent[1],extent[2]}(Int64.(-2*ones(Int64, extent)))
     walkmap = BitArray{2}(trues(extent))
     plan = Vector{Vector{Tuple}}(undef, nb_robots)
     frontiers = Set()
@@ -30,7 +30,7 @@ function initialize_model(
     nb_obstacles = [nb_obstacles]
 
     properties = (
-        seen_all_gridmap = MVector{nb_robots, MMatrix}(MMatrix{extent[1],extent[2]}(Int8.(zeros(Int8, extent))) for i in 1:nb_robots),
+        seen_all_gridmap = MVector{nb_robots, MMatrix}(MMatrix{extent[1],extent[2]}(Int64.(zeros(Int64, extent))) for i in 1:nb_robots),
         nb_obstacles, 
         invisible_cells,
         nb_robots
