+++ src/Utils.jl
@@ -3,7 +3,7 @@ function compute_actions_cenMCTS(nb_robots)
     theta = [i*pi/4 for i in 0:7]
     rad_actions = [(round(cos(θ),digits=2),round(sin(θ), digits=2)) for θ in theta]
 
-    ids = collect(Int8(i) for i in 1:nb_robots)
+    ids = collect(Int64(i) for i in 1:nb_robots)
     
     actionsNR = [Vector{Action}(undef,length(rad_actions)) for i in 1:nb_robots]
